Support playback & seeking in media files over 2G.
BUG=95805
TEST=unittests of touched files pass, manually tested w/ sonyh2.webm which is >5G (internal only), and trybots.


Review URL: http://codereview.chromium.org/8349009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@106183 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/glue/multipart_response_delegate.h
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2006-2009 The Chromium Authors. All rights reserved.
+// Copyright (c) 2011 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 //
@@ -92,9 +92,9 @@ class MultipartResponseDelegate {
   // in a multipart response.
   // Returns true on success.
   static bool ReadContentRanges(const WebKit::WebURLResponse& response,
-                                int* content_range_lower_bound,
-                                int* content_range_upper_bound,
-                                int* content_range_instance_size);
+                                int64* content_range_lower_bound,
+                                int64* content_range_upper_bound,
+                                int64* content_range_instance_size);
 
  private:
   friend class MultipartResponseDelegateTester;  // For unittests.
